Standardize interface tasks
Remove operation-specific tasks and replace with object tasks, w/ op
parameter. This makes the tasks easier to maintain as well.

Eventually what should probably happen is a single Ruby library should
be created for dealing with the data, and tasks should just call methods
in it (after including the library file as a dependency)