Rename nodedata fields
Use "puppet_environment" instead of environment.
Use "puppet_classes" instead of classes.

In user interfaces, we will be recommending calling this data Puppet
Environment, and Puppet Classes. The Puppet prefix is to avoid collision
with internal use of terms like environment. For consistency with that
UI recommendation, the database fields should also be prefixed with
puppet_*.